feat!: Introduce configuration file
There are a lot of command line options and arguments for this tool,
making the user experience less than ideal. To simplify usage, we are
introducing a configuration file.

Each configuration file will include a named profile which can be
referenced via the `--profile` command line option. If one isn't
specified, we fall back to `default`.

Entries in the configuration file can be overridden using the equivalent
command options (e.g. `---bastion-user`, `--bastion-port`).