build: Add release process
This commit introduces a cargo-make step called release. Using
cocogitto, the application versions will be bumped, a changelog
generated, and a new tag pushed to GitHub.

Upon tag push, a GitHub action workflow will build static binaries of
this application and upload them as build artifacts to the newly created
tag. It will also use the changelog entry that was just generated to set
the release notes for that tag.